Cleanup in sweep_boundary module
