+++ modules/LinearBoltzmannSolvers/A_LBSSolver/lbs_solver.cc
@@ -11,6 +11,10 @@
 #include "Groupset/lbs_groupset.h"
 #include "mesh/MeshHandler/chi_meshhandler.h"
 #include "mesh/MeshContinuum/chi_meshcontinuum.h"
+#include "mesh/SweepUtilities/SweepBoundary/boundary_reflecting.h"
+#include "mesh/SweepUtilities/SweepBoundary/boundary_vacuum.h"
+#include "mesh/SweepUtilities/SweepBoundary/boundary_iso_homo.h"
+#include "mesh/SweepUtilities/SweepBoundary/boundary_aniso_hetero.h"
 #include "math/TimeIntegrations/time_integration.h"
 #include "math/SpatialDiscretization/FiniteElement/PiecewiseLinear/PieceWiseLinearDiscontinuous.h"
 #include "physics/PhysicsMaterial/chi_physicsmaterial.h"
